refactor: 💡 Split navigation stacks & tabs into new components
