+++ navigation/stacks/AuthStacks.tsx
@@ -1,0 +1,21 @@
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { SignInScreen, SignUpScreen } from "../../screens";
+
+const Stack = createNativeStackNavigator();
+
+const AuthStacks = (): JSX.Element => {
+  return (
+    <Stack.Navigator
+      initialRouteName="SignIn"
+      screenOptions={{
+        contentStyle: { backgroundColor: "#fffbfe" }, // NOTE: React Native Paper background colour theme
+        headerShown: false,
+      }}
+    >
+      <Stack.Screen name="SignUp" component={SignUpScreen} />
+      <Stack.Screen name="SignIn" component={SignInScreen} />
+    </Stack.Navigator>
+  );
+};
+
+export default AuthStacks;
